web/requirements.txt
flask
validator_collection
pymongo
gunicorn
python-dotenv
requests
